requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pandas==2.2.3
pydantic-settings==2.8.1
